hotfix: fixing some field types

+++ MoviesSita.Domain/Entities/Movies.cs
@@ -8,17 +8,17 @@ namespace MoviesSita.Domain.Entities
 {
     public class Movies
     {
-        public int? id { get; set; }
+        public long? id { get; set; }
         public string? title { get; set; } 
         public float? vote_average { get; set; }
-        public int? vote_count { get; set; }
+        public long? vote_count { get; set; }
         public string? status { get; set; }
         public string? release_date { get; set; }
-        public int? revenue { get; set; }
-        public int? runtime { get; set; }
+        public long? revenue { get; set; }
+        public long? runtime { get; set; }
         public bool?  adult { get; set; }
         public string? backdrop_path { get; set; }
-        public int? budget { get; set; }
+        public long? budget { get; set; }
         public string? homepage { get; set; }
         public string? imdb_id { get; set; }
         public string? original_language { get; set; }
